handle no results case, add reset button
